+++ packages/terra-markdown/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Added
 * Added documentation on required peerDependencies
 
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
+
 2.28.0 - (June 28, 2019)
 ------------------
 ### Changed